fix(tokens): derive every manifest's token namespaces from one shared reader
CI caught `@lostgradient/chat`'s components.json as stale: the previous commit
changed that package's generator to read namespaces from the registry but never
regenerated the artifact. Regenerating it surfaced the larger problem — three
packages publish a components.json, each hand-listed the same array, and I had
only fixed two generators. `packages/editor` still advertised `color`, a
namespace CIN-33 deleted.

Rather than write the same reader a third time, the helper moves to
`packages/components/scripts/token-namespaces.ts`, which the chat and editor
generators already import other shared pieces from. One derivation means the
three manifests cannot disagree with the corpus or with each other.

All three now report the same 29 namespaces with `color` absent.
