+++ packages/chat/scripts/generate-component-artifacts.ts
@@ -1,4 +1,4 @@
-import { existsSync, readFileSync as nodeReadFileSync } from 'node:fs';
+import { existsSync } from 'node:fs';
 import { readdir } from 'node:fs/promises';
 import { basename, join } from 'node:path';
 
@@ -21,6 +21,7 @@ import {
   type ComponentMetadata,
 } from '../../components/scripts/generate-component-metadata.ts';
 import { discoverDirectoryComponents } from '../../components/scripts/generate-exports.ts';
+import { readTokenNamespaces } from '../../components/scripts/token-namespaces.ts';
 import { categories, statusLevels } from '../../components/src/manifest.meta.ts';
 
 function isRecord(value: unknown): value is Record<string, unknown> {
@@ -48,49 +49,6 @@ type ManifestComponent = Omit<ComponentMetadata, 'isExperimental'> & {
 
 const PACKAGE_ROOT = join(import.meta.dir, '..');
 
-/** One field of an already-narrowed object, without asserting a shape over it. */
-function readOwnField(source: object, field: string): unknown {
-  return Object.hasOwn(source, field)
-    ? (Object.getOwnPropertyDescriptor(source, field)?.value as unknown)
-    : undefined;
-}
-
-/**
- * The top-level token namespaces the corpus actually declares, read from the
- * generated registry.
- *
- * Hand-listing them put a stale array into every published `components.json`:
- * it named `color`, which CIN-33 deleted, and omitted namespaces the corpus had
- * gained since. Deriving it means the manifest cannot disagree with the tokens
- * it describes.
- */
-function readTokenNamespaces(): string[] {
-  const registryPath = join(
-    PACKAGE_ROOT,
-    '..',
-    'components',
-    'src',
-    'tokens',
-    'registry.generated.json',
-  );
-  const parsed: unknown = JSON.parse(nodeReadFileSync(registryPath, 'utf8'));
-  if (typeof parsed !== 'object' || parsed === null) {
-    throw new Error(`${registryPath} is not a JSON object.`);
-  }
-  const entries = readOwnField(parsed, 'entries');
-  if (!Array.isArray(entries)) {
-    throw new Error(`${registryPath} has no \`entries\` array.`);
-  }
-  const namespaces = new Set<string>();
-  for (const entry of entries) {
-    if (typeof entry !== 'object' || entry === null) continue;
-    const path = readOwnField(entry, 'path');
-    if (typeof path !== 'string' || path === '') continue;
-    const [namespace] = path.split('.');
-    if (namespace !== undefined) namespaces.add(namespace);
-  }
-  return [...namespaces].sort();
-}
 const PACKAGES_ROOT = join(PACKAGE_ROOT, '..');
 const COMPONENTS_ROOT = join(PACKAGE_ROOT, 'src', 'lib', 'components');
 const PLAYGROUND_EXAMPLES_ROOT = join(PACKAGES_ROOT, 'playground', 'src', 'examples');
