+++ packages/editor/components.json
@@ -8,7 +8,37 @@
     "frameworkVersionRange": ">=5.56.0 <6",
     "classPrefix": "cinder-",
     "cssVarPrefix": "--cinder-",
-    "tokenNamespaces": ["color", "space", "radius", "ring", "type", "motion", "shadow"],
+    "tokenNamespaces": [
+      "accent",
+      "border",
+      "button",
+      "chart",
+      "checker",
+      "content",
+      "control",
+      "ease",
+      "fill",
+      "font",
+      "gap",
+      "leading",
+      "motion",
+      "opacity",
+      "overlay",
+      "pad",
+      "radius",
+      "ring",
+      "scrollbar",
+      "shadow",
+      "space",
+      "status",
+      "surface",
+      "text",
+      "toggle",
+      "touch",
+      "tracking",
+      "type",
+      "z"
+    ],
     "stylesEntry": "@lostgradient/cinder/styles",
     "schemaDialect": "https://json-schema.org/draft/2020-12/schema"
   },
